requirements.txt
Flask-SQLAlchemy==3.1.1
Flask==3.0.2
Flask-Login==0.6.3
python-dotenv==1.0.1
Flask-Uploads==0.2.1